remove useFileUpload and shift logic back to form using useController
